move regexp integration tests to new file
